changed strategy for deleting stuff generated by autosummary
Instead of running Sphinx on a temporary copy of the docs, have autosummary
generate its files inside a subdirectory of doc/ and remove that. This has the
advantage of not needing to copy stuff around but, OTOH, the drawback that the
name of autosummary's target folder needs to be kept synchronised in three
places, and the generated files end up between the user's own source files.